requirements.txt
gradio>=4.0.0
requests>=2.31.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
